requirements.txt
fastapi
uvicorn
celery
redis
python-dotenv
moviepy
Pillow
jinja2
